+++ crates/wdl-engine/src/backend/docker.rs
@@ -34,6 +34,8 @@ use super::PullResults;
 use super::TaskExecutionBackend;
 use super::TaskExecutionConstraints;
 use super::TaskExecutionResult;
+#[cfg(unix)]
+use crate::CLEANUP_TASK_NAME_PREFIX;
 use crate::CancellationContext;
 use crate::EvaluationPath;
 use crate::Events;
@@ -452,8 +454,6 @@ pub struct DockerBackend {
     max_memory: u64,
     /// The task manager for the backend.
     manager: TaskManager,
-    /// The name generator for tasks.
-    names: Arc<Mutex<GeneratorIterator<UniqueAlphanumeric>>>,
 }
 
 impl DockerBackend {
@@ -517,7 +517,6 @@ impl DockerBackend {
             max_cpu,
             max_memory,
             manager,
-            names,
         })
     }
 }
@@ -654,23 +653,14 @@ impl TaskExecutionBackend for DockerBackend {
                 .ok_or_else(|| anyhow::anyhow!("{results}"))?;
             let container = container.clone();
 
-            let name = format!(
-                "{id}-{generated}",
-                id = request.id,
-                generated = self
-                    .names
-                    .lock()
-                    .expect("generator should always acquire")
-                    .next()
-                    .expect("generator should never be exhausted")
-            );
+            let task_name = request.name.to_string();
 
             let task = DockerTask {
                 config: self.config.clone(),
                 request,
                 container,
                 backend: self.inner.clone(),
-                name,
+                name: task_name.clone(),
                 max_cpu,
                 max_memory,
                 gpu,
@@ -682,15 +672,7 @@ impl TaskExecutionBackend for DockerBackend {
                     // The task completed, perform cleanup on unix platforms
                     #[cfg(unix)]
                     {
-                        let name = format!(
-                            "docker-chown-{id}",
-                            id = self
-                                .names
-                                .lock()
-                                .expect("generator should always acquire")
-                                .next()
-                                .expect("generator should never be exhausted")
-                        );
+                        let name = format!("{CLEANUP_TASK_NAME_PREFIX}chown-{task_name}");
 
                         let task = CleanupTask {
                             name,